# k8s/networking/04-api-gateway-openapi.yaml
...
swagger: "2.0"    # Google Cloud API Gateway uses OpenAPI 2.0 (Swagger), not 3.0
info:
  title: "Mistral-7B Inference API"
  description: |
    Production inference gateway for Mistral-7B-Instruct-v0.3.
    OpenAI-compatible API hosted on GKE with vLLM backend.
    Requires a valid API key in the x-api-key header or as a query parameter.
  version: "0.1.0"
  contact:
    name: "Platform Team"
    email: "platform@example.com"
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0"

# The base URL for all paths. Google Cloud API Gateway will use the HTTPS
# managed endpoint — consumers call this domain.
# Replace PROJECT_ID with your GCP project ID.
host: "mistral-gateway.apigateway.PROJECT_ID.cloud.goog"
basePath: "/"
schemes:
  - https   # API Gateway always serves HTTPS — no HTTP option

# ── Security definition ───────────────────────────────────────────────────────
# WHY api_key security?
#   API key authentication is the correct choice for service-to-service calls:
#     • No expiry (unlike JWT access tokens that expire in 1 hour)
#     • No refresh flow needed (simplifies client code)
#     • Per-key rate limiting and quota enforcement at API Gateway level
#     • Easy revocation: delete the key in GCP Console → instant effect
#
# HOW IT WORKS:
#   Client sends: GET /v1/models  with header  x-api-key: AIzaSy...
#   API Gateway validates the key against GCP API key service
#   If valid → forward to GKE Ingress backend
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "x-api-key"   # Header name clients must include
    in: "header"        # Can also be "query" (x-api-key=... in URL) but header is safer

# Apply API key requirement to ALL paths by default
security:
  - api_key: []

# ── Backend configuration ─────────────────────────────────────────────────────
# WHY x-google-backend?
#   This is a Google Cloud API Gateway extension that specifies where to forward
#   requests after authentication and validation.
#   address: the GKE Ingress external IP (or DNS name).
#
#   Replace INGRESS_EXTERNAL_IP with the IP from:
#     kubectl get ingress mistral-ingress -n mistral-gateway -o jsonpath='{.status.loadBalancer.ingress[0].ip}'
#
#   WHY protocol: h2?
#     HTTP/2 (h2) multiplexes multiple requests over a single TCP connection.
#     This is more efficient for long-running inference requests (stream responses).
#     HTTP/1.1 requires a new connection per request — higher overhead.
...
paths:
...
  /v1/chat/completions:
    post:
      operationId: "createChatCompletion"
      summary: "Create a chat completion"
      description: |
        OpenAI-compatible chat completions endpoint.
        Proxied to vLLM running Mistral-7B-Instruct-v0.3.
        Maximum response time: 120 seconds (VLLM_TIMEOUT_SECONDS).
      tags:
        - inference
      security:
        - api_key: []
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - model
              - messages
            properties:
              model:
                type: string
                description: "Must be 'mistralai/Mistral-7B-Instruct-v0.3'"
                example: "mistralai/Mistral-7B-Instruct-v0.3"
              messages:
                type: array
                description: "Conversation history in OpenAI message format"
                items:
                  type: object
                  required:
                    - role
                    - content
                  properties:
                    role:
                      type: string
                      enum: [system, user, assistant]
                    content:
                      type: string
              max_tokens:
                type: integer
                description: "Max tokens to generate. Ceiling enforced by FastAPI: 4096."
                default: 512
                minimum: 1
                maximum: 4096
              temperature:
                type: number
                description: "Sampling temperature. 0 = deterministic."
                default: 0.7
                minimum: 0
                maximum: 2
              stream:
                type: boolean
                description: "If true, response is streamed as Server-Sent Events."
                default: false
      responses:
        200:
          description: "Chat completion response"
          schema:
            type: object
            properties:
              id:
                type: string
              object:
                type: string
                example: "chat.completion"
              created:
                type: integer
              model:
                type: string
              choices:
                type: array
                items:
                  type: object
                  properties:
                    index:
                      type: integer
                    message:
                      type: object
                    finish_reason:
                      type: string
              usage:
                type: object
                properties:
                  prompt_tokens:
                    type: integer
                  completion_tokens:
                    type: integer
                  total_tokens:
                    type: integer
        400:
          description: "Bad request (missing required fields, max_tokens exceeded ceiling)"
        401:
          description: "Invalid or missing API key"
        429:
          description: "Rate limit exceeded — try again after the Retry-After header interval"
        504:
          description: "Gateway timeout — vLLM inference exceeded 120 seconds"